+++ NEWS.md
@@ -13,6 +13,7 @@ Replace this text with you changelog entry. Look at the examples below.
 - [#1032] Refactor BaseRequest callbacks into configurable lambdas
 - [#1040] Clear mixins from ActiveRecord DSL and save only overridable API. It
   allows to use this mixins in Doorkeeper ORM extensions with minimum code boilerplate.
+- [#907] Fix lookup for matching tokens in certain edge-cases
 
 ## 4.3.0